fix: исправила m и ms
